9차 - 게시물 유저 연동 및 게시물 수정 삭제 권한 처리

+++ src/main/java/com/dgm/board/service/PostService.java
@@ -1,6 +1,9 @@
 package com.dgm.board.service;
 
 import com.dgm.board.exception.post.PostNotFoundException;
+import com.dgm.board.exception.user.UserAlreadyExistsException;
+import com.dgm.board.exception.user.UserNotAllowedException;
+import com.dgm.board.model.entity.UserEntity;
 import com.dgm.board.model.post.Post;
 import com.dgm.board.model.post.PostPatchRequestBody;
 import com.dgm.board.model.post.PostPostRequestBody;
@@ -39,22 +42,26 @@ public class PostService {
     }
 
 
-        public Post createPost(PostPostRequestBody postPostRequestBody) {
+        public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
 
-                PostEntity postEntity = new PostEntity();
-                postEntity.setBody(postPostRequestBody.getBody());
-                PostEntity savedPostEntity = postEntityRepository.save(postEntity);
+                PostEntity postEntity = postEntityRepository.save(PostEntity.of(postPostRequestBody.getBody(), currentUser));
 
-                return Post.from(savedPostEntity);
-
+                return Post.from(postEntity);
         }
 
-        public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody) {
+        public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody, UserEntity currentUser) {
+
+            System.out.println("Updating Post: Current User -> " + currentUser.getUserId());
+            System.out.println("Updating Post: Current User -> " + currentUser.getUsername());
 
                 PostEntity postEntity = postEntityRepository.findById(postId)
                         .orElseThrow(
                                 () -> new PostNotFoundException(postId)
                         );
+
+                if (!postEntity.getUser().equals(currentUser)) {
+                    throw new UserNotAllowedException();
+                }
 
                 postEntity.setBody(postPatchRequestBody.getBody());
                 PostEntity updatedPostEntity = postEntityRepository.save(postEntity);
@@ -63,12 +70,16 @@ public class PostService {
 
         }
 
-        public void deletePost(Long postId) {
+        public void deletePost(Long postId, UserEntity currentUser) {
 
                 PostEntity postEntity = postEntityRepository.findById(postId)
                         .orElseThrow(
                                 () -> new PostNotFoundException(postId)
                         );
+
+                if (!postEntity.getUser().equals(currentUser)) {
+                    throw new UserNotAllowedException();
+                }
 
                 postEntityRepository.delete(postEntity);
